refactor: 게스트 관련 dto를 별도 패키지로 분리
